backend/requirements.txt
channels-redis==3.4.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0
favicon==0.7.0
gunicorn==20.1.0
httptools==0.4.0
mysqlclient==1.4.6
pillow==9.1.1
pip-chill==1.0.1
pyopenssl==22.0.0
python-decouple==3.6
python-dotenv==0.20.0
pyyaml==6.0
service-identity==21.1.0
uvicorn==0.18.2
watchfiles==0.15.0
websockets==10.3
